refactor warnings command
